style: Reduce padding between quick menu toggle text and separator
The padding between the text and the arrow button separator effectively
was 2 * base_padding, while the spacing between the icon and the text
was just 1.5 * base_padding. This was a bit uneven and also could result
in the text getting ellipsized when it wasn't necessary.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/6432
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2685>