style: Enforce a min spacing of the dash to the left/right screen edges
The dash currently doesn't enforce any spacing when it comes to the left and
right edges of the monitor. With the right amount of icons, it will come
very close to or even touch the edge.

Make sure this "edge case" also looks good by adding a padding to the left
and right of the container.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3278>